refactoring draw pins

+++ middleware/include/middleware/gscheme/editor/gs_draw.h
@@ -53,7 +53,7 @@ private:
     // key = nodeIndex
     std::vector<uint8_t> m_actualNodes;
     // key = nodeIndex
-    std::vector<float> m_maxOutputPinNameWidthPerNode;
+    std::vector<uint32_t> m_maxOutputPinNameWidthPerNode;
 
 // persistent data
 private:
